# api/openapi/developer.yaml
openapi: 3.1.0
info:
  title: Qeet ID — Developer API (keys, webhooks, secrets, agents)
  version: 0.2.0
  description: 'Qeet ID identity platform — full HTTP surface.


    This document is generated to cover EVERY route mounted by the chi router (internal/http/router.go) and is guarded by internal/http/openapi_coverage_test.go, which fails CI if any mounted route is undocumented. Launch-critical families (Auth, OIDC/OAuth, SCIM, SAML, Users, Tenants, RBAC, API Keys) carry fully-described request/response bodies; remaining admin CRUD routes carry accurate path/method/params/security with summarized bodies (noted in each operation''s description).


    Auth modes: `bearerAuth` (user/service JWT), `apiKeyAuth` (`Authorization: ApiKey <key>`), `scimBearer` (per-tenant SCIM token), `clientCredentials` (OAuth M2M), and `ssoCookie` (hosted-login SSO session).'
servers:
  - url: http://localhost:4001
    description: Local (docker compose)
  - url: http://localhost:4000
    description: Local (make dev-backend)
...
paths:
...
  /v1/agents/token:
    post:
      operationId: post_v1_agents_token
      summary: Mint agent token
      description: >-
        An AI agent authenticates with its agent_id + secret (JSON body or HTTP Basic) and receives an ephemeral access token marked actor_type="agent" (+ agent_id claim), verifiable on the standard JWKS path. No user session.
      tags:
        - Agents
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
                - secret
              properties:
                agent_id:
                  type: string
                secret:
                  type: string
      responses:
        '200':
          description: Short-lived agent access token
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/tenants/{tenantID}/agents:
    get:
      operationId: get_v1_tenants_ByTenantID_agents
      summary: List agents
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Agents (secrets are write-only, not returned)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: post_v1_tenants_ByTenantID_agents
      summary: Create agent
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - sponsor_user_id
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                token_ttl_seconds:
                  type: integer
                sponsor_user_id:
                  type: string
                  format: uuid
                  description: The named human owner accountable for this agent. Must be a member of this tenant.
      responses:
        '201':
          description: Agent created (secret returned once)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/tenants/{tenantID}/agents/{id}:
    delete:
      operationId: delete_v1_tenants_ByTenantID_agents_ById
      summary: Delete agent
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patch_v1_tenants_ByTenantID_agents_ById
      summary: Update agent
      description: >-
        Sets or clears the agent's disabled_at. A disabled agent can no longer
        mint tokens; existing tokens remain valid until they expire.
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - disabled
              properties:
                disabled:
                  type: boolean
      responses:
        '204':
          description: Updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/tenants/{tenantID}/agents/{id}/suspend:
    post:
      operationId: post_v1_tenants_ByTenantID_agents_ById_suspend
      summary: Suspend agent
      description: Move an agent to `suspended` (reversible). Its tokens are denied at authz within one request cycle.
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: New lifecycle status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tenants/{tenantID}/agents/{id}/resume:
    post:
      operationId: post_v1_tenants_ByTenantID_agents_ById_resume
      summary: Resume agent
      description: Move a suspended agent back to `active`.
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: New lifecycle status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tenants/{tenantID}/agents/{id}/decommission:
    post:
      operationId: post_v1_tenants_ByTenantID_agents_ById_decommission
      summary: Decommission agent
      description: Permanently retire an agent (terminal). It can no longer mint tokens and is excluded from listings.
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: New lifecycle status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tenants/{tenantID}/agents/kill-all:
    post:
      operationId: post_v1_tenants_ByTenantID_agents_kill_all
      summary: Suspend all agents
      description: >-
        Security incident-response control. Sets disabled_at on all of the
        tenant's agents in a single transaction. Returns the number suspended.
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Agents suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  suspended:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/tenants/{tenantID}/agents/sponsored-by/{userID}:
    get:
      operationId: get_v1_tenants_ByTenantID_agents_sponsored_by_ByUserID
      summary: List agents sponsored by a user
      description: What an admin needs to see before offboarding userID — every agent they're the accountable owner of.
      tags:
        - Agents
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
        - name: userID
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User or service JWT: `Authorization: Bearer <token>`.'
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'API key credential. Send `Authorization: ApiKey <key>` (note the `ApiKey ` prefix). Accepted on the same authenticated routes as bearerAuth.'
    scimBearer:
      type: http
      scheme: bearer
      description: 'Per-tenant SCIM provisioning token presented by the IdP: `Authorization: Bearer <scim-token>`. Distinct from user JWTs; rotated via the SCIM admin endpoints.'
    clientCredentials:
      type: oauth2
      description: OAuth 2.0 client credentials. Clients authenticate at the token endpoint via form params (`client_id`/`client_secret`) or HTTP Basic. Revocation/introspection use the same client auth.
      flows:
        clientCredentials:
          tokenUrl: /v1/oauth/token
          scopes: {}
    ssoCookie:
      type: apiKey
      in: cookie
      name: qe_ls
      description: Hosted-login SSO session cookie (HttpOnly). Set by POST /v1/auth/session and consumed by the OAuth authorize/consent flow.
...